requirements.txt
streamlit>=1.32.0
cobra>=0.29.0
pandas>=2.0.0
plotly>=5.20.0
optlang>=1.8.1
numpy>=1.26.0
lxml>=5.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9
requests>=2.31.0
httpx>=0.27.0
biopython>=1.83
